+++ src/systems/linux.sh
@@ -151,14 +151,14 @@ add-linux-kconfig(revision, architecture=x86, lkc_binding_file=) {
     if [[ -n $lkc_binding_file ]]; then
         add-kconfig-model \
             --system linux \
-            --revision "${revision}[$architecture]" \
+            --revision "$(revision-with-context "$revision" "$architecture")" \
             --kconfig-file "$kconfig_file" \
             --lkc-binding-file "$lkc_binding_file" \
             --environment "$environment"
     else
         add-kconfig \
             --system linux \
-            --revision "${revision}[$architecture]" \
+            --revision "$(revision-with-context "$revision" "$architecture")" \
             --kconfig-file "$kconfig_file" \
             --lkc-directory scripts/kconfig \
             --environment "$environment"